requirements.txt


# ====================== LangChain 大语言模型应用框架 ======================

langchain>=1.0.7
langchain-openai>=1.0.3
langchain-community>=0.4.1
langchain-core>=1.0.7

# ====================== Web 服务框架与服务器 ======================
fastapi>=0.100.0
uvicorn>=0.23.0

# ====================== 数据库驱动 ======================

mysql-connector-python~=9.4.0


# ====================== 向量数据库与云服务 ======================
pinecone~=7.3.0
dashscope>=1.14.0

# ====================== 工具与辅助库 ======================
python-dotenv>=1.0.0
requests>=2.31.0
pydantic~=2.11.7


# ====================== python格式化工具 ======================
black==26.3.1
